requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pycparser==2.21
pygame==2.5.2
pyparsing==3.1.1
pyreqs==0.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
screeninfo==0.8.1
seaborn==0.13.0
sh==2.0.6
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.196
urllib3==2.0.6
